[SPOT-280][REFACTOR] Organize Schedule Domain
