+++ src/main/java/com/example/spot/refactor/member/domain/Member.java
@@ -3,14 +3,14 @@ package com.example.spot.refactor.member.domain;
 import com.example.spot.legacy.domain.*;
 import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyComment;
 import com.example.spot.refactor.study.domain.aggregate.studypost.LikedStudyPost;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyQuiz;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuiz;
 import com.example.spot.refactor.member.domain.association.StudyJoinReason;
 import com.example.spot.refactor.common.entity.BaseEntity;
 import com.example.spot.refactor.member.domain.enums.Carrier;
 import com.example.spot.refactor.member.domain.enums.Gender;
 import com.example.spot.refactor.member.domain.enums.LoginType;
 import com.example.spot.refactor.member.domain.enums.Status;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyQuizSubmission;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuizSubmission;
 import com.example.spot.legacy.domain.mapping.MemberScrap;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.member.domain.association.MemberTheme;
@@ -19,7 +19,7 @@ import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVote;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVoteParticipant;
 import com.example.spot.refactor.member.domain.association.PreferredRegion;
 import com.example.spot.refactor.member.domain.association.PreferredStudy;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudySchedule;
+import com.example.spot.refactor.schedule.domain.StudySchedule;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPost;
 import com.example.spot.refactor.study.domain.aggregate.studypost.StudyPostComment;
 import com.example.spot.refactor.member.presentation.dto.MemberRequestDTO.MemberUpdateDTO;
