+++ src/main/java/com/example/spot/refactor/schedule/domain/aggregate/StudyQuiz.java
@@ -1,7 +1,8 @@
-package com.example.spot.refactor.study.domain.aggregate.studyschedule;
+package com.example.spot.refactor.schedule.domain.aggregate;
 
 import com.example.spot.refactor.member.domain.Member;
 
+import com.example.spot.refactor.schedule.domain.StudySchedule;
 import jakarta.persistence.*;
 import lombok.*;
