+++ src/main/java/com/example/spot/refactor/schedule/domain/enums/StudySchedulePeriod.java
@@ -1,4 +1,4 @@
-package com.example.spot.refactor.study.domain.enums;
+package com.example.spot.refactor.schedule.domain.enums;
 
 public enum StudySchedulePeriod {
     NONE, DAILY, WEEKLY, BIWEEKLY, MONTHLY
