+++ src/main/java/com/example/spot/refactor/schedule/domain/repository/StudyScheduleRepository.java
@@ -1,5 +1,6 @@
-package com.example.spot.refactor.study.domain.aggregate.studyschedule;
+package com.example.spot.refactor.schedule.domain.repository;
 
+import com.example.spot.refactor.schedule.domain.StudySchedule;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
