+++ src/main/java/com/example/spot/refactor/schedule/domain/validation/validator/ExistStudyScheduleValidator.java
@@ -1,8 +1,8 @@
-package com.example.spot.refactor.study.domain.validation.validator;
+package com.example.spot.refactor.schedule.domain.validation.validator;
 
 import com.example.spot.refactor.common.api.code.status.ErrorStatus;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyScheduleRepository;
-import com.example.spot.refactor.study.domain.validation.annotation.ExistStudySchedule;
+import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
+import com.example.spot.refactor.schedule.domain.validation.annotation.ExistStudySchedule;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
