+++ src/main/java/com/example/spot/refactor/study/application/MemberStudyCommandServiceImpl.java
@@ -7,13 +7,17 @@ import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
 import com.example.spot.legacy.domain.MemberReport;
 import com.example.spot.legacy.domain.Notification;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.*;
+import com.example.spot.refactor.schedule.domain.*;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuiz;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuizSubmission;
+import com.example.spot.refactor.schedule.domain.repository.StudyQuizRepository;
+import com.example.spot.refactor.schedule.domain.repository.StudyQuizSubmissionRepository;
+import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
 import com.example.spot.refactor.study.domain.aggregate.studytodo.StudyToDo;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.legacy.domain.enums.NotifyType;
 import com.example.spot.refactor.member.domain.enums.Status;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyQuizSubmission;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
 import com.example.spot.refactor.study.domain.aggregate.studyvote.StudyVoteParticipant;
 import com.example.spot.refactor.study.domain.aggregate.Study;
