+++ src/main/java/com/example/spot/refactor/study/presentation/dto/request/ScheduleRequestDTO.java
@@ -1,6 +1,6 @@
 package com.example.spot.refactor.study.presentation.dto.request;
 
-import com.example.spot.refactor.study.domain.enums.StudySchedulePeriod;
+import com.example.spot.refactor.schedule.domain.enums.StudySchedulePeriod;
 import com.example.spot.legacy.validation.annotation.TextLength;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
