+++ src/test/java/com/example/spot/service/study/studyschedule/StudyAttendanceQueryServiceTest.java
@@ -2,11 +2,15 @@ package com.example.spot.service.study.studyschedule;
 
 import com.example.spot.refactor.common.api.exception.handler.StudyHandler;
 import com.example.spot.refactor.member.domain.Member;
+import com.example.spot.refactor.schedule.domain.*;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuiz;
+import com.example.spot.refactor.schedule.domain.aggregate.StudyQuizSubmission;
+import com.example.spot.refactor.schedule.domain.repository.StudyQuizRepository;
+import com.example.spot.refactor.schedule.domain.repository.StudyQuizSubmissionRepository;
+import com.example.spot.refactor.schedule.domain.repository.StudyScheduleRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMember;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.*;
 import com.example.spot.refactor.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.refactor.member.domain.enums.Gender;
-import com.example.spot.refactor.study.domain.aggregate.studyschedule.StudyQuizSubmission;
 import com.example.spot.refactor.study.domain.aggregate.Study;
 import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.StudyMemberRepository;
